+++ includes/functions/coloring.h
@@ -1,31 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   minirt.h                                           :+:      :+:    :+:   */
+/*   coloring.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/03/08 20:57:46 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/16 20:18:19 by jvacaris         ###   ########.fr       */
+/*   Created: 2022/03/16 20:01:43 by jvacaris          #+#    #+#             */
+/*   Updated: 2022/03/16 20:19:24 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef MINIRT_H
-# define MINIRT_H
-# include <unistd.h>
-# include <stdlib.h>
-# include <string.h>
-# include <limits.h>
-# include <stdio.h>
-# include <fcntl.h>
-# include <mlx.h>
-# include <math.h>
-// # include "libft/libft.h"
-# include "libft.h"
-# include "constants.h"
+#ifndef COLORING_H
+# define COLORING_H
 # include "dtos.h"
-# include "functions/coloring.h"
-# include "functions/parser.h"
-# include "functions/utils.h"
+
+t_colors	**set_color_matrix(t_itemlist *items, t_coords **v_matrix);
 
 #endif